requirements.txt
langchain_community
langchain_core
streamlit
langchain_groq
python-dotenv